companies/requirements.txt
pandas
pyyaml
numpy
pyarrow
yfinance
playwright
